排查Attention算子获取整形属性GetAttrPointer潜在风险fix GetAttrPointerInt potential risk



# message auto-generated for no-merge-commit merge:
!2980 merge master into master

排查Attention算子获取整形属性GetAttrPointer潜在风险fix GetAttrPointerInt potential risk

Created-by: tangkaidi
Commit-by: tangkaidi
Merged-by: cann-robot
Description: ## 描述
排查与修改Attention算子获取整形属性GetAttrPointer潜在风险。统一改用GetAttrPointer<int64_t>来获取算子整型属性。1. 避免属性值超过int32_t时的截断风险。2. 算子原型定义中的Int匹配的是int64_t，GetAttrPointer<T>中的模板参数需要与算子原型中的类型一一对应避免未定义行为。
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #000-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] 🐛 Bug 修复
- [ ] ✨ 新特性
- [ ] ⚡ 性能优化
- [ ] ♻️ 重构
- [ ] 🧪 测试
- [ ] 📦 构建/CI
- [ ] 🔧 配置变更
- [ ] 📝 文档更新
- [ ] ⬆️ 依赖升级
- [ ] 🔒 安全修复
- [ ] 🧹 代码清理
- [ ] ❓ 其他，请描述：



See merge request: cann/ops-transformer!2980

+++ attention/attention_update/op_host/decode_update_tiling.cpp
@@ -66,8 +66,8 @@ ge::graphStatus DecodeUpdateTiling(gert::TilingContext *context) {
     OP_LOGD(nodeName, "Tiling initing");
 
     auto attrs = context->GetAttrs();
-    const uint32_t* updateTypePtr = attrs->GetAttrPointer<uint32_t>(ATTR_UPDATETYPE_INDEX);
-    const uint32_t* spPtr = attrs->GetAttrPointer<uint32_t>(ATTR_SP_INDEX);
+    const int64_t* updateTypePtr = attrs->GetAttrPointer<int64_t>(ATTR_UPDATETYPE_INDEX);
+    const int64_t* spPtr = attrs->GetAttrPointer<int64_t>(ATTR_SP_INDEX);
 
     auto updateType = *updateTypePtr;
     auto sp = *spPtr;
@@ -84,7 +84,7 @@ ge::graphStatus DecodeUpdateTiling(gert::TilingContext *context) {
     const uint32_t totalLength = lseShape.GetDim(LSE_TOTAL_LENGTH_DIM);
     const uint32_t hd = inShape.GetDim(IN_HD_DIM);
 
-    OP_LOGD(nodeName, "TotalLength of b*s*hc is %u, sp is %u, hd is %u",
+    OP_LOGD(nodeName, "TotalLength of b*s*hc is %u, sp is %ld, hd is %u",
                                             totalLength, sp, hd);
     
     uint32_t blockDims = 0;
